Simplify ipr::Decl a bit (#78)
At some point in the design of the IPR, we decided that just about any expression should be subject to parameterization -- not just declarations, but also types (which later gave rise to template aliases, that CWG unwisely renamed to alias templates) and classic expressions.  A corollary of that is that every declaration was potentially the result of an instantiation.

This patch starts a series where we take a slightly different approach to implementing that idea by representing request for specialization as distinct expression nodes.  First we remove the remnant of the previous implementation tactic.