Predicate for testing global region (#163)
This patch adds a `Region::global()` predicate that holds for the global region, i.e. the region of the global scope.

Fix #33 .

As a side note, ensure that the tests are run with flags consistent with the build.  Someone more versed in CMake should have a look and devise a better way to do this so we don't have two places to update.