Add Tor type for object constructors and destructors
Standard C++ does not assign types to constructor or destructor declarations.  The IPR until now has relied on platform ABI-dependent interpretations of those special functions as regular functions with some ABI-dependent return type.  A high level representation of C++ is expected to abstract over those differences.  This is the role of this new type `Tor` class.  It is essentially a `Function`, minus the `target()` type.