Remove `ipr::Decl::position`
The operation `Decl::position()` was to enable indexing into a scope based on position.  That model breaks down when a scope is made of pieces coming from various scopes (e.g. import of modules) in contemporary C++.  It does not seem to have lived up to its promises.  Removed thusly.