Region has a body
This patch adds the `body()` operation to a `ipr::Region`: it returns the sequence of generalized expressions (classic expressions, statements, directives, declarations, etc.) that makes up a lexical region.

As a consequence, a function and user-defined types all have bodies: the sequence of expressions lexically written in their defining initializers.