The type of an alias is now always that of its initializer (#208)
In general, the type of an initializer is that of its initializer -- especially for type aliases.  However, structured bindings can introduce aliases too (at object level) and the corresponding binding may have acquired a _cv-qualifier_.  Therefore, this patch removes the incorrect override in the interface of `Alias` -- since it is overriden later in the implementation, it is never executed.

In short, this is a removal of a wrong dead code.