Simplify Tor type (#212)
While in general, a `Tor` type is close to a `Function` type and, in principle, should have a language linkage, that does not actually happen in practice.  This patch removes that possibility for now.  It should be reinstated when we have more evidence of its usefulness (non-`C` language linkage constructors) in practice.