Upgrade representation of exception specification
Back in 2004, only types could be mentioned in function exception specification.  That ability was later named _dynamic exception specification_.   The IPR representation at the time opted to integrate that directly in the type system even when it wasn't at the ISO C++ level.  Throwing anything was modeled by the ellipsis type `...`.

Very late in the C++11 development process, `noexcept` was introduced for specifying that a function could be thrown something or not (without mentioning what the type of that thing could be).  While the operand (at the source level) can be an arbitrary expression, when fully evaluated, the value of the operand be only be `false` or `true`.

This patch adds ability to specify `noexcept`-ness in the type system.

Fixes #150