Remove `name_factory::get_type_id`
A `Type_id` name is unique by the virtue of being associated with a unique type.  We don't need a separate uniquing factory function.  Plus, uses of `Type_id` names is mostly internal, for technical reason.  That factory function is gone with this patch.  That exposes further simplification opportunities of the existing implementation codebase - in preparation for making linkage part of all types.