Remove `Rname` (#227)
Back in 2004, we introduced `Rname` as a mathematical device to handle the situation where function declarations and template declarations are matched/unified without regard to the names of the parameters.  With the move to differentiating between core expressions and located expressions, it is no longer necessary to have `Rname` around.  Removed with this patch.  I took the opportunity to further simplify the implementation codebase.  This is a source-level breaking change.