Fix #114 (#229)
That bug report observed that it is no longer possible -- after the `Name` refactoring -- to build a `Template_id` where the template declaration is designated by an expression.  Something that was previously possible thanks to the original simpler design for `Name` as part of the `Expr` hierarchy.  Fixed with this patch.

Fixes #114