Make `impl::singleton_overload` contain its declset (#235)
The general model of a scope is that it is a sequence of declarations.  The declarations are further divided into overload sets.  Each overload set contains a declset : all the declarations in that scope for the same name and same type.

The existing implementation  `impl::unique_decl` contains `singleton_overload` which is logically backward.  Fixed with this patch.  Now, a homogeneous scope is a sequence of `singleton_overload` s.  Each `singleton_overload` contains a decl(set).