Fix #23 (#236)
An exception parameter declaration for an EH block acts functionally like a function parameter, and a variable.  But, unlike function parameters or variables, they are not lexically initialized -- their initializations are performed by the runtime at the places where an exception is thrown.  There are arguments for reusing `ipr::Parameter` or `ipr::Var` to represent that.  This patch removes that ambiguity by introducing a distinct node class to represent EH parameters.  Most of the code is about capturing the nesting structure of EH handlers, region, and scope.

Fixes #23