Add `ipr::Phased_evaluation` (#273)
Both _asm-declaration_ and _static\_assert-declaration_ are now uniformly represented as phased evaluation of general expressions `ipr::Asm` and `ipr::Static_assert`.

This unified description comes at the cost of an indirection from a `Directive` to an `Expr`.  However, it allows for other type expressions (e.g. parameterized _prolongations_) to be described the same.  Furthermore, an `if constexpr` and an `if consteval` are now just phased evaluation of ordinary `if-statement`.  Similarly, static initialization can also be described as a phased evaluation.